Address CodeRabbit review feedback
- LayerPanel: the post-save metadata merge reads the store's current
  layer metadata instead of the click-time closure, so concurrent
  metadata updates (auto-refresh, time-slider binding) are not clobbered
  by a slow write/re-read round trip.
- postgis write: the failure log records the scrubbed error message
  instead of a full traceback, which could embed the raw DSN and break
  the module's no-credentials-in-logs guarantee.
- Moved the saved-PostgreSQL-connection helpers (and their constants)
  from the add-data UI folder to lib/saved-postgres-connections.ts so
  lib/postgis-connections.ts no longer depends on a UI feature folder;
  add-data/helpers re-exports them for its existing importers.
